Add details to manifest validation/display
Rename command to View instead.
