+++ src/Commands/ListCommand.cs
@@ -1,6 +1,5 @@
 ﻿using System.ComponentModel;
 using System.Diagnostics;
-using System.Text.Json;
 using DotNetConfig;
 using Spectre.Console;
 using Spectre.Console.Cli;
